refactor: move MESSAGE_TYPES and text utils out of chat.channels
MESSAGE_TYPES moves to apps.channels.const; URL_REGEX, MARKDOWN_REF_PATTERN
and strip_urls_and_emojis move to apps.channels.text_utils. This removes
channels_v2's dependency on the legacy apps.chat.channels module.
