+++ apps/channels/channels_v2/api_channel.py
@@ -21,7 +21,7 @@ from apps.channels.channels_v2.stages.core import (
     SessionResolutionStage,
 )
 from apps.channels.channels_v2.stages.terminal import ActivityTrackingStage, PersistenceStage
-from apps.chat.channels import MESSAGE_TYPES
+from apps.channels.const import MESSAGE_TYPES
 from apps.chat.exceptions import ChannelException
 from apps.chat.models import Chat
 from apps.experiments.models import Experiment, Participant, SessionStatus
