+++ apps/channels/channels_v2/connect_channel.py
@@ -8,7 +8,7 @@ from apps.channels.channels_v2.channel_base import ChannelBase
 from apps.channels.channels_v2.pipeline import MessageProcessingContext
 from apps.channels.channels_v2.sender import ChannelSender
 from apps.channels.clients.connect_client import CommCareConnectClient
-from apps.chat.channels import MESSAGE_TYPES
+from apps.channels.const import MESSAGE_TYPES
 from apps.chat.exceptions import ChannelException
 
 logger = logging.getLogger("ocs.channels")
