+++ apps/channels/channels_v2/stages/terminal.py
@@ -10,7 +10,7 @@ from apps.annotations.models import TagCategories
 from apps.channels.channels_v2.pipeline import MessageProcessingContext
 from apps.channels.channels_v2.stages.base import ProcessingStage
 from apps.channels.channels_v2.stages.core import RESET_COMMAND
-from apps.chat.channels import MESSAGE_TYPES
+from apps.channels.const import MESSAGE_TYPES
 from apps.chat.models import ChatMessage, ChatMessageType
 from apps.files.models import File, FilePurpose
 from apps.ocs_notifications.notifications import (
