+++ apps/channels/channels_v2/telegram_channel.py
@@ -17,8 +17,8 @@ from apps.channels.channels_v2.channel_base import ChannelBase
 from apps.channels.channels_v2.pipeline import MessageProcessingContext
 from apps.channels.channels_v2.sender import ChannelSender
 from apps.channels.channels_v2.stages.terminal import DeliveryErrorHandler, MessageDeliveryFailure
+from apps.channels.const import MESSAGE_TYPES
 from apps.channels.datamodels import TelegramMessage
-from apps.chat.channels import MESSAGE_TYPES
 from apps.experiments.models import ParticipantData
 from apps.service_providers.file_limits import can_send_on_telegram
