+++ apps/channels/tests/channels/concrete/test_slack_channel.py
@@ -3,7 +3,7 @@ from unittest.mock import MagicMock
 import pytest
 
 from apps.channels.channels_v2.slack_channel import SlackChannel
-from apps.chat.channels import MESSAGE_TYPES
+from apps.channels.const import MESSAGE_TYPES
 from apps.chat.exceptions import ChannelException
